Split util files into separate files for clarity
