+++ src/cli/init/doPostInitActions.ts
@@ -1,0 +1,23 @@
+import { execSync } from "child_process";
+
+export const doPostInitActions = (dirName: string): void => {
+  // run npm install from the new directory
+  process.chdir(dirName);
+  execSync("npm install", {
+    stdio: "inherit",
+  });
+
+  // setup git
+  execSync("git init --initial-branch=main", {
+    stdio: "inherit",
+  });
+
+  // try to open vscode
+  try {
+    execSync("code .", {
+      stdio: "inherit",
+    });
+  } catch {
+    console.warn("VSCode was not found, IDE will not automatically open.");
+  }
+};
